project/requirements.txt
cs50
Flask
Flask-Session
werkzeug
python-dotenv
